move everything into chat provider
